requirements.txt
ptvsd
canal-python